refactor(protocol): move reconcileBatches into the package
