+++ cmd/evener-hub/frontend/src/protocol/README.md
@@ -7,20 +7,21 @@ types, wire errors with their session classifiers, the rejection classifier
 and the user-facing message helpers every failure display goes through, the
 pure question formatter, the ask_user question parser and the answered recap
 it reads back out of a transcript, the live-question derivation an answering
-dock renders from a thread, the translation that turns a composer's
-`[image N]` attachment markers into prose at send, the thread view model
-and its notification reducer, the activity tree parser, merge and disclosure
-rules, the job log tail parser, the send/queue availability table, the
-send/steer/queue/drain routing decisions a composer makes off it, the stable
-delegate status rule, the slash invocation and catalog visibility rules the
-palette and composer share, the display formatters both apps render counts,
-durations and clock times with, the text and argument helpers a tool call's
-rendering is built from, and the doc-pane URL builders, which hang their hrefs
-off a base origin the host supplies (empty for a same-origin web page). The
-doc-pane data layer is published at the `./docContent` subpath as well, where
-`readDocFile` takes the host's `DocPort` - that base origin paired with a
-fetch: the package issues no request of its own and names neither an origin
-nor a credentials policy.
+dock renders from a thread, the batch reconciliation that keeps an in-flight
+answer's questions frozen while late ones arrive, the translation that turns a
+composer's `[image N]` attachment markers into prose at send, the thread view
+model and its notification reducer, the activity tree parser, merge and
+disclosure rules, the job log tail parser, the send/queue availability table,
+the send/steer/queue/drain routing decisions a composer makes off it, the
+stable delegate status rule, the slash invocation and catalog visibility rules
+the palette and composer share, the display formatters both apps render
+counts, durations and clock times with, the text and argument helpers a tool
+call's rendering is built from, and the doc-pane URL builders, which hang
+their hrefs off a base origin the host supplies (empty for a same-origin web
+page). The doc-pane data layer is published at the `./docContent` subpath as
+well, where `readDocFile` takes the host's `DocPort` - that base origin paired
+with a fetch: the package issues no request of its own and names neither an
+origin nor a credentials policy.
 
 Build and qualify from this directory with `npm run qualification`. The runner
 packs the package, installs that tarball into a temporary consumer, and then,
