+++ cmd/evener-hub/frontend/src/protocol/tsconfig.build.json
@@ -21,6 +21,7 @@
     "askAnswers.ts",
     "askShared.ts",
     "deriveAskQuestions.ts",
+    "reconcileBatches.ts",
     "attachmentMarkers.ts",
     "activityData.ts",
     "activityRows.ts",
